Simplified Async-FastAPI & React Industrial Binance Futures Terminal/trading_bot/requirements.txt
fastapi
uvicorn
aiohttp
python-dotenv
rich
questionary
pydantic
